feat: 后端 Maven→Gradle 迁移 + 接口规范化 + Swagger;前端令牌纪律 + E2E 回归门 + 修复
后端(对齐 cn.csg.datapod 规范,走公共仓库不引内网)
- Maven→Gradle 多工程(删 pom + settings/build.gradle + gradlew;bootJar 仍落 target/,容器零改)
- 统一出参 Result/PageResult/ResponseCode + 全局 BusinessException/GlobalExceptionHandler
- 分页入参统一 PageQuery;全 Controller @Validated + 自定义体 @Valid
- 集成 Swagger/OpenAPI(springdoc 2.3.0,OpenApiConfig + Result/PageResult/PageQuery @Schema;/swagger-ui、/v3/api-docs)
- 确权变更基线反查上一版真实确权(表3行 ConfirmTableItem + 权益卡片 + 认定意见)
- 逐权制卡(多权拼接拆成一卡一权,授权侧精确命中);系统→业务域映射逐表带出
- 批量授权可授资源池接口(/dpr/auth/grantable/open-filter)

前端
- 批量授权一站式:唯一确权目录入口 GrantableCatalogTree(同源 fullCatalogTree)+ 资源池过滤(先确后授/权属可授/经营权对外开放)+ 表5/表6 字段(联系人·联系方式·跨系统域自动判定·业务域逐表)
- 设计令牌纪律:全站灰阶/语义色令牌化 + scripts/lint-tokens.mjs 守卫(prebuild 门禁)
- 修复 lib/chartBase.js initChart 返回 undefined(ECharts setOption 返回 void)致全景/统计两页整页崩 + 下钻失效
- E2E 回归门:playwright.config + e2e/*(route-load 40 路由 + 批量授权流程),npm run e2e

测试:后端 gradle test 311 绿;前端 build 绿 + E2E 41/41;线上状态机链路(提交→审批→制卡/发证/驳回)实跑通过

+++ prm-backend/dpr-confirm-service/src/main/java/com/csg/prm/confirm/integration/AssetCardArchiveService.java
@@ -49,7 +49,7 @@ public class AssetCardArchiveService {
         long size = query.getSize() <= 0 ? 10 : query.getSize();
         int from = (int) Math.min((current - 1) * size, rows.size());
         int to = (int) Math.min(from + size, rows.size());
-        return new PageResult<>(rows.size(), current, size, rows.subList(from, to));
+        return new PageResult<>((long) rows.size(), (int) current, (int) size, rows.subList(from, to));
     }
 
     /**
